Fix bogus trap on remote_owner check
Some call sites pass a uint8_t to objnum_remote_to_local, which is
zero-extended up to an int.  The check for owner==-1 then fails, causing
the sanity check to trap to the debugger, even though the situation is
normal and harmless.  Switch the type to int8_t to ensure that the value
is not sign/zero-extended.

+++ common/main/multi.h
@@ -230,8 +230,8 @@ struct owned_remote_objnum
 
 extern int GetMyNetRanking();
 extern void ClipRank (ubyte *rank);
-objnum_t objnum_remote_to_local(int remote_obj, int owner);
-short objnum_local_to_remote(objnum_t local_obj, sbyte *owner);
+objnum_t objnum_remote_to_local(uint16_t remote_obj, int8_t owner);
+uint16_t objnum_local_to_remote(objnum_t local_obj, int8_t *owner);
 owned_remote_objnum objnum_local_to_remote(objnum_t local);
 void map_objnum_local_to_remote(int local, int remote, int owner);
 void map_objnum_local_to_local(objnum_t objnum);
